refactor : 컴포넌트 > App 컴포넌트에서 분리
