+++ .eslintrc.js
@@ -3,5 +3,12 @@ module.exports = {
   extends: '@react-native',
   rules: {
     'react/react-in-jsx-scope': 'off',
+    'react-native/no-inline-styles': 0,
+    'prettier/prettier': [
+      'error',
+      {
+        'no-inline-styles': false,
+      },
+    ],
   },
 };
