+++ constants/index.js
@@ -1,1 +1,1 @@
-export const SOURCE_URL = 'https://gloddy.vercel.app';
+export const SOURCE_URL = 'http://localhost:3000';
